feat: add react hook to use predeployed accounts
