+++ .changeset/clean-tables-turn.md
@@ -1,0 +1,5 @@
+---
+"@dojoengine/predeployed-connector": patch
+---
+
+feat: add react hook to get predeployed accounts
