+++ .gitignore
@@ -7,3 +7,4 @@ docs
 lerna-debug.log
 
 .turbo/
+**/**/vite.config.ts.timestamp-*
