+++ examples/example-vite-kitchen-sink/src/components/starknet-provider.tsx
@@ -1,20 +1,16 @@
 import type { PropsWithChildren } from "react";
 import { mainnet } from "@starknet-react/chains";
 import { jsonRpcProvider, StarknetConfig, voyager } from "@starknet-react/core";
-import { env, getRpcUrl } from "@/env";
-import {
-    predeployedAccounts,
-    PredeployedAccountsConnector,
-} from "@dojoengine/predeployed-connector";
-
-let pa: PredeployedAccountsConnector[] = [];
-predeployedAccounts({
-    rpc: env.VITE_CONTROLLER_RPC,
-    id: "katana",
-    name: "Katana",
-}).then((p) => (pa = p));
+import { getRpcUrl } from "@/env";
+import { usePredeployedAccounts } from "@dojoengine/predeployed-connector/react";
+import { dojoConfig } from "../../dojoConfig";
 
 export default function StarknetProvider({ children }: PropsWithChildren) {
+    const { connectors } = usePredeployedAccounts({
+        rpc: dojoConfig.rpcUrl as string,
+        id: "katana",
+        name: "Katana",
+    });
     const provider = jsonRpcProvider({
         rpc: () => ({ nodeUrl: getRpcUrl() }),
     });
@@ -23,7 +19,7 @@ export default function StarknetProvider({ children }: PropsWithChildren) {
         <StarknetConfig
             chains={[mainnet]}
             provider={provider}
-            connectors={[...pa]}
+            connectors={[...connectors]}
             explorer={voyager}
             autoConnect
         >
