+++ examples/example-vite-react-app-recs/package.json
@@ -26,7 +26,6 @@
         "react-dom": "^18.3.1",
         "rxjs": "^7.8.1",
         "starknet": "catalog:",
-        "vite-plugin-top-level-await": "^1.4.4",
         "vite-plugin-wasm": "^3.3.0"
     },
     "devDependencies": {
@@ -43,6 +42,7 @@
         "postcss": "^8.4.47",
         "tailwindcss": "^3.4.13",
         "typescript": "^5.6.2",
-        "vite": "^4.5.5"
+        "vite": "^4.5.5",
+        "vite-plugin-top-level-await": "^1.4.4"
     }
 }
