+++ examples/example-vite-react-sdk/package.json
@@ -28,7 +28,6 @@
         "react-dom": "catalog:",
         "starknet": "catalog:",
         "uuid": "^10.0.0",
-        "vite-plugin-top-level-await": "^1.5.0",
         "vite-plugin-wasm": "^3.4.1",
         "zustand": "^4.5.6"
     },
@@ -46,6 +45,7 @@
         "tailwindcss": "^3.4.17",
         "typescript": "^5.7.3",
         "typescript-eslint": "^8.24.1",
-        "vite": "^5.4.14"
+        "vite": "^5.4.14",
+        "vite-plugin-top-level-await": "^1.5.0"
     }
 }
