+++ examples/example-vite-react-sql/src/components/starknet-provider.tsx
@@ -1,29 +1,18 @@
-import { useEffect, useState, type PropsWithChildren } from "react";
+import { type PropsWithChildren } from "react";
 import { mainnet } from "@starknet-react/chains";
 import { jsonRpcProvider, StarknetConfig, voyager } from "@starknet-react/core";
 import { dojoConfig } from "../../dojoConfig";
-import {
-    predeployedAccounts,
-    PredeployedAccountsConnector,
-} from "@dojoengine/predeployed-connector";
+import { usePredeployedAccounts } from "@dojoengine/predeployed-connector/react";
 
 export default function StarknetProvider({ children }: PropsWithChildren) {
-    const [connectors, setConnectors] = useState<
-        PredeployedAccountsConnector[]
-    >([]);
+    const { connectors } = usePredeployedAccounts({
+        rpc: dojoConfig.rpcUrl as string,
+        id: "katana",
+        name: "Katana",
+    });
     const provider = jsonRpcProvider({
         rpc: () => ({ nodeUrl: dojoConfig.rpcUrl as string }),
     });
-
-    useEffect(() => {
-        if (connectors.length === 0) {
-            predeployedAccounts({
-                rpc: dojoConfig.rpcUrl as string,
-                id: "katana",
-                name: "Katana",
-            }).then(setConnectors);
-        }
-    }, [connectors]);
 
     return (
         <StarknetConfig
