+++ examples/example-vite-react-threejs-recs/package.json
@@ -54,7 +54,6 @@
         "tailwindcss-animate": "^1.0.7",
         "three": "^0.160.1",
         "vite-plugin-svgr": "^4.2.0",
-        "vite-plugin-top-level-await": "^1.4.4",
         "vite-plugin-wasm": "^3.3.0",
         "zustand": "^4.5.5"
     },
@@ -79,6 +78,7 @@
         "eslint-plugin-storybook": "^0.6.15",
         "storybook": "^7.6.20",
         "typescript": "^5.6.2",
-        "vite": "^4.5.5"
+        "vite": "^4.5.5",
+        "vite-plugin-top-level-await": "^1.4.4"
     }
 }
