+++ examples/example-vue-app-recs/package.json
@@ -22,7 +22,6 @@
         "@dojoengine/utils": "workspace:*",
         "@latticexyz/utils": "^2.2.8",
         "starknet": "catalog:",
-        "vite-plugin-top-level-await": "^1.4.4",
         "vite-plugin-wasm": "^3.3.0",
         "vue": "^3.5.8"
     },
@@ -30,6 +29,7 @@
         "@vitejs/plugin-vue": "^5.1.4",
         "typescript": "^5.6.2",
         "vite": "^5.4.7",
+        "vite-plugin-top-level-await": "^1.4.4",
         "vue-tsc": "^2.1.6"
     }
 }
