+++ packages/predeployed-connector/package.json
@@ -21,11 +21,17 @@
             "import": "./dist/index.js",
             "types": "./dist/index.d.ts"
         },
+        "./react": {
+            "import": "./dist/src/react.js",
+            "types": "./dist/src/react.d.ts"
+        },
         "./package.json": "./package.json"
     },
     "devDependencies": {
         "@rollup/plugin-commonjs": "^28.0.0",
         "@vitest/coverage-v8": "^1.6.0",
+        "@types/react": "catalog:",
+        "@types/react-dom": "catalog:",
         "eslint": "^8.57.1",
         "prettier": "^2.8.8",
         "tsup": "^8.3.0",
@@ -35,7 +41,9 @@
     },
     "peerDependencies": {
         "@starknet-react/core": "catalog:",
-        "starknet": "catalog:"
+        "starknet": "catalog:",
+        "react": "catalog:",
+        "react-dom": "catalog:"
     },
     "repository": {
         "type": "git",
