+++ packages/predeployed-connector/tsup.config.ts
@@ -4,4 +4,8 @@ import { tsupConfig } from "../../tsup.config";
 
 export default defineConfig({
     ...(tsupConfig as Options),
+    entry: {
+        index: "src/index.ts",
+        "src/react": "src/react/index.ts",
+    },
 });
